renamed package expetion to exception
