+++ executor-service/src/main/java/executor/service/exception/ConstructorNotFoundException.java
@@ -1,4 +1,4 @@
-package executor.service.expetion;
+package executor.service.exception;
 
 public class ConstructorNotFoundException extends RuntimeException {
