+++ executor-service/src/main/java/executor/service/exception/FoundSecondImplementationException.java
@@ -1,4 +1,4 @@
-package executor.service.expetion;
+package executor.service.exception;
 
 public class FoundSecondImplementationException extends RuntimeException {
     public FoundSecondImplementationException(Class<?> interfaceClazz){
