+++ executor-service/src/main/java/executor/service/exception/InstanceCreationException.java
@@ -1,4 +1,4 @@
-package executor.service.expetion;
+package executor.service.exception;
 
 public class InstanceCreationException extends RuntimeException{
