+++ executor-service/src/main/java/executor/service/exception/MethodInvocationException.java
@@ -1,4 +1,4 @@
-package executor.service.expetion;
+package executor.service.exception;
 
 public class MethodInvocationException extends RuntimeException {
