+++ executor-service/src/main/java/executor/service/exception/NoImplementationsFoundException.java
@@ -1,4 +1,4 @@
-package executor.service.expetion;
+package executor.service.exception;
 
 public class NoImplementationsFoundException extends RuntimeException{
